fix(test-layout): move set_delivering_activated test to own file (1000-line gate)
tests.rs crossed the 1000-line file-size gate after the new
set_delivering_activated_preserves_receiving_flag test was added in
f95923b. Moved it to streaming_event_flag_tests.rs matching the
existing per-topic split pattern (audit_tests.rs, metrics_tests.rs,
etc.). tests.rs now 978 lines.

No test logic changes — pure relocation plus the mod declaration in
db/mod.rs.
