+++ crates/rs-core/src/db/mod.rs
@@ -53,6 +53,9 @@ mod metrics_tests;
 #[cfg(test)]
 mod pool_tests;
 
+#[cfg(test)]
+mod streaming_event_flag_tests;
+
 /// Create a SQLite connection pool.
 pub async fn create_pool(db_path: &Path) -> Result<SqlitePool> {
     let url = format!("sqlite:{}?mode=rwc", db_path.display());
